updates EventList component (refactor)
