+++ packages/read-and-create-calendar-events/frontend/react/src/EventList.jsx
@@ -1,81 +1,82 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import './styles/calendar.scss';
-import {
-  getSevenDaysFromTodayDateTimestamp,
-  getTodaysDateTimestamp,
-} from './utils/date';
+// import {
+//   getSevenDaysFromTodayDateTimestamp,
+//   getTodaysDateTimestamp,
+// } from './utils/date';
 import EventPreview from './EventPreview';
 import { initializeScrollShadow, handleScrollShadows } from './utils/calendar';
 
 function EventList({
-  serverBaseUrl,
-  userId,
-  calendarId,
+  // serverBaseUrl,
+  // userId,
+  // calendarId,
   setSelectedEvent,
   selectedEvent,
   isLoading,
-  setIsLoading,
+  // setIsLoading,
+  events,
 }) {
-  const [calendarEvents, setCalendarEvents] = useState([]);
+  // const [calendarEvents, setCalendarEvents] = useState([]);
   const [showTopScrollShadow, setShowTopScrollShadow] = useState(false);
   const [showBottomScrollShadow, setShowBottomScrollShadow] = useState(false);
 
   // let loading = true;
   console.log(isLoading);
 
-  useEffect(() => {
-    setIsLoading(true);
-    const getCalendarEvents = async () => {
-      if (calendarId) {
-        // loading = true;
+  // useEffect(() => {
+  //   setIsLoading(true);
+  //   const getCalendarEvents = async () => {
+  //     if (calendarId) {
+  //       // loading = true;
 
-        try {
-          const startsAfter = getTodaysDateTimestamp(); // today
-          const endsBefore = getSevenDaysFromTodayDateTimestamp(); // 7 days from today
+  //       try {
+  //         const startsAfter = getTodaysDateTimestamp(); // today
+  //         const endsBefore = getSevenDaysFromTodayDateTimestamp(); // 7 days from today
 
-          const queryParams = new URLSearchParams({
-            limit: 5,
-            startsAfter,
-            endsBefore,
-            calendarId,
-          });
+  //         const queryParams = new URLSearchParams({
+  //           limit: 5,
+  //           startsAfter,
+  //           endsBefore,
+  //           calendarId,
+  //         });
 
-          const url = `${serverBaseUrl}/nylas/read-events?${queryParams.toString()}`;
+  //         const url = `${serverBaseUrl}/nylas/read-events?${queryParams.toString()}`;
 
-          const res = await fetch(url, {
-            method: 'GET',
-            headers: {
-              Authorization: userId,
-              'Content-Type': 'application/json',
-            },
-            params: {
-              calendarId,
-            },
-          });
+  //         const res = await fetch(url, {
+  //           method: 'GET',
+  //           headers: {
+  //             Authorization: userId,
+  //             'Content-Type': 'application/json',
+  //           },
+  //           params: {
+  //             calendarId,
+  //           },
+  //         });
 
-          if (!res.ok) {
-            throw new Error(res.statusText);
-          }
+  //         if (!res.ok) {
+  //           throw new Error(res.statusText);
+  //         }
 
-          const data = await res.json();
+  //         const data = await res.json();
 
-          setCalendarEvents(data);
+  //         setCalendarEvents(data);
 
-          // loading = false;
-          setIsLoading(false);
-        } catch (err) {
-          console.warn(`Error reading calendar events:`, err);
-        }
-      }
-    };
+  //         // loading = false;
+  //         setIsLoading(false);
+  //       } catch (err) {
+  //         console.warn(`Error reading calendar events:`, err);
+  //       }
+  //     }
+  //   };
 
-    getCalendarEvents();
-  }, [serverBaseUrl, userId, calendarId]);
+  //   getCalendarEvents();
+  // }, [serverBaseUrl, userId, calendarId]);
 
   useEffect(() => {
     initializeScrollShadow('.event-list-container', setShowBottomScrollShadow);
-  }, [calendarEvents]);
+  }, [events]);
 
   useEffect(() => {
     window.addEventListener('resize', () =>
@@ -101,11 +102,11 @@ function EventList({
       <div
         className={`scroll-shadow top${showTopScrollShadow ? '' : ' hidden'}`}
       ></div>
-      {calendarEvents.length === 0 ? (
-        <p>{loading ? 'Loading events.' : 'No events scheduled.'}</p>
+      {events.length === 0 ? (
+        <p>{isLoading ? 'Loading events.' : 'No events scheduled.'}</p>
       ) : (
         <ul className="event-list">
-          {calendarEvents.map((calendarEvent) => (
+          {events.map((calendarEvent) => (
             <div
               key={calendarEvent.id}
               onClick={() => handleEventSelect(calendarEvent)}
@@ -128,13 +129,14 @@ function EventList({
 }
 
 EventList.propTypes = {
-  serverBaseUrl: PropTypes.string.isRequired,
-  userId: PropTypes.string.isRequired,
-  calendarId: PropTypes.string,
+  // serverBaseUrl: PropTypes.string.isRequired,
+  // userId: PropTypes.string.isRequired,
+  // calendarId: PropTypes.string,
   setSelectedEvent: PropTypes.func,
   selectedEvent: PropTypes.object,
   isLoading: PropTypes.bool.isRequired,
-  setIsLoading: PropTypes.func.isRequired,
+  // setIsLoading: PropTypes.func.isRequired,
+  events: PropTypes.array.isRequired,
 };
 
 export default EventList;
